ignore MIME types that throw errors during display (fixes #776), split display into separate file
