+++ src/IJulia.jl
@@ -353,6 +353,10 @@ include("hmac.jl")
 include("eventloop.jl")
 include("stdio.jl")
 include("msg.jl")
+include("display.jl")
+include("magics.jl")
+include("comm_manager.jl")
+include("execute_request.jl")
 include("handlers.jl")
 include("heartbeat.jl")
 include("inline.jl")
